create steamers module
